[PURPLE-151] 평가 집계 로직 적용
