+++ src/main/java/com/pikachu/purple/application/review/port/in/reviewevaluation/CreateReviewEvaluationUseCase.java
@@ -1,0 +1,16 @@
+package com.pikachu.purple.application.review.port.in.reviewevaluation;
+
+import com.pikachu.purple.bootstrap.review.vo.EvaluationFieldVO;
+import com.pikachu.purple.domain.review.Review;
+import java.util.List;
+
+public interface CreateReviewEvaluationUseCase {
+
+    void invoke(Command command);
+
+    record Command(
+        Review review,
+        List<EvaluationFieldVO> evaluationFieldVOs
+    ) {}
+
+}
